+++ packages/components/src/components/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultCardContent } from './DefaultCardContent.vue'