+++ packages/data-explorer/src/mappers/utils.ts
@@ -1,0 +1,3 @@
+import { MetaDataAttribute } from '@/types/ApiResponse'
+
+export const getCategoricals = (attributes: MetaDataAttribute[]): MetaDataAttribute[] => attributes.filter(item => item.fieldType.includes('CATEGORICAL'))