+++ packages/data-explorer/src/store/actions.ts
@@ -5,7 +5,7 @@ import { tryAction } from './helpers'
 import * as metaDataRepository from '@/repository/metaDataRepository'
 import * as dataRepository from '@/repository/dataRepository'
 import * as metaDataService from '@/repository/metaDataService'
-import * as metaFilterMapper from '@/repository/metaFilterMapper'
+import * as metaFilterMapper from '@/mappers/metaFilterMapper'
 
 export default {
   getTableSettings: tryAction(async ({ commit, state }: { commit: any, state: ApplicationState },
@@ -32,6 +32,7 @@ export default {
 
     const rsqlQuery = getters.filterRsql
 
+    commit('setTableData', [])
     if (isCustomCard) {
       columns = state.tableSettings.customCardAttrs.split(',').map(attribute => attribute.trim())
       tableData = await dataRepository.getTableDataDeepReference(state.tableName, metaData, columns, rsqlQuery)
@@ -39,8 +40,10 @@ export default {
       columns = metaDataService.getAttributesfromMeta(metaData).splice(0, state.tableSettings.collapseLimit)
       tableData = await dataRepository.getTableDataWithLabel(state.tableName, metaData, columns, rsqlQuery)
     }
-
-    commit('setTableData', tableData)
+    if (getters.filterRsql === rsqlQuery) {
+      // retrieved results are still relevant
+      commit('setTableData', tableData)
+    }
   },
   fetchTableViewData: async ({ commit, getters }: { commit: any, getters: any }, payload: {tableName: string}) => {
     const tableName = payload.tableName
@@ -49,8 +52,12 @@ export default {
 
     const rsqlQuery = getters.filterRsql
 
+    commit('setTableData', [])
     const tableData = await dataRepository.getTableDataWithLabel(tableName, metaData, metaDataService.getAttributesfromMeta(metaData), rsqlQuery)
-    commit('setTableData', tableData)
+    if (getters.filterRsql === rsqlQuery) {
+      // retrieved results are still relevant
+      commit('setTableData', tableData)
+    }
   },
   // expanded default card
   fetchRowDataLabels: async ({ commit, state, getters }: { commit: any, state: ApplicationState, getters: any }, payload: {rowId: string}) => {
@@ -60,7 +67,11 @@ export default {
     const metaData = await metaDataRepository.fetchMetaData(state.tableName)
     commit('setMetaData', metaData)
     const rsqlQuery = getters.filterRsql
-    const rowData = await dataRepository.getRowDataWithReferenceLabels(state.tableName, payload.rowId, metaData, rsqlQuery)
-    commit('updateRowData', { rowId: payload.rowId, rowData })
+    commit('updateRowData', [])
+    const rowData = await dataRepository.getRowDataWithReferenceLabels(state.tableName, payload.rowId, metaData)
+    if (getters.filterRsql === rsqlQuery) {
+      // retrieved results are still relevant
+      commit('updateRowData', { rowId: payload.rowId, rowData })
+    }
   }
 }