+++ packages/data-explorer/src/types/ApplicationState.ts
@@ -24,6 +24,10 @@ export type TableSetting = {
   isShop: boolean,
 }
 
+export type FilterSelections = {
+  [s: string]: string | string[]
+}
+
 export type FilterDefinition = {
   name: string,
   label: string,
@@ -51,6 +55,6 @@ export default interface ApplicationState {
     hideSidebar: boolean
     definition: FilterDefinition[]
     shown: string[]
-    selections: StringMap
+    selections: FilterSelections
   }
 }